+++ loan-mate/src/main/java/com/fisa/bank/account/application/model/spending/CategorySpending.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.account.application.model;
+package com.fisa.bank.account.application.model.spending;
 
 import java.math.BigDecimal;
