+++ loan-mate/src/main/java/com/fisa/bank/account/application/service/helper/UserAccountContextService.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.account.application.service;
+package com.fisa.bank.account.application.service.helper;
 
 import lombok.RequiredArgsConstructor;
