+++ loan-mate/src/main/java/com/fisa/bank/account/application/usecase/GetMonthlySpendingUseCase.java
@@ -1,6 +1,6 @@
 package com.fisa.bank.account.application.usecase;
 
-import com.fisa.bank.account.application.model.MonthlySpending;
+import com.fisa.bank.account.application.model.spending.MonthlySpending;
 
 public interface GetMonthlySpendingUseCase {
   MonthlySpending execute(Long accountId, int year, int month);
