+++ loan-mate/src/main/java/com/fisa/bank/account/application/usecase/GetUserSpendingLimitUseCase.java
@@ -2,7 +2,7 @@ package com.fisa.bank.account.application.usecase;
 
 import java.util.Optional;
 
-import com.fisa.bank.account.application.model.UserSpendingLimit;
+import com.fisa.bank.account.application.model.spending.UserSpendingLimit;
 
 public interface GetUserSpendingLimitUseCase {
