+++ loan-mate/src/main/java/com/fisa/bank/account/persistence/repository/impl/AccountRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.account.persistence.repository;
+package com.fisa.bank.account.persistence.repository.impl;
 
 import lombok.RequiredArgsConstructor;
 
@@ -7,6 +7,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Repository;
 
 import com.fisa.bank.account.application.repository.AccountRepository;
+import com.fisa.bank.account.persistence.repository.jpa.JpaAccountRepository;
 import com.fisa.bank.persistence.account.entity.Account;
 import com.fisa.bank.persistence.user.entity.User;
