+++ loan-mate/src/main/java/com/fisa/bank/account/persistence/repository/impl/SpendingRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.account.persistence.repository;
+package com.fisa.bank.account.persistence.repository.impl;
 
 import lombok.RequiredArgsConstructor;
 
@@ -12,6 +12,9 @@ import java.util.Map;
 import org.springframework.stereotype.Repository;
 
 import com.fisa.bank.account.application.repository.SpendingRepository;
+import com.fisa.bank.account.persistence.repository.CategoryAmount;
+import com.fisa.bank.account.persistence.repository.jpa.JpaAccountTransactionRepository;
+import com.fisa.bank.account.persistence.repository.jpa.JpaCardTransactionRepository;
 import com.fisa.bank.persistence.account.entity.id.AccountId;
 import com.fisa.bank.persistence.account.enums.ConsumptionCategory;
