+++ loan-mate/src/main/java/com/fisa/bank/account/persistence/repository/jpa/JpaAccountTransactionRepository.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.account.persistence.repository;
+package com.fisa.bank.account.persistence.repository.jpa;
 
 import java.math.BigDecimal;
