+++ loan-mate/src/main/java/com/fisa/bank/account/persistence/repository/jpa/JpaCardTransactionRepository.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.account.persistence.repository;
+package com.fisa.bank.account.persistence.repository.jpa;
 
 import java.util.List;
 
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import com.fisa.bank.account.persistence.repository.CategoryAmount;
 import com.fisa.bank.persistence.account.entity.CardTransaction;
 import com.fisa.bank.persistence.account.entity.id.AccountId;
