+++ loan-mate/src/main/java/com/fisa/bank/account/persistence/repository/jpa/JpaUserSpendingLimitRepository.java
@@ -1,4 +1,4 @@
-package com.fisa.bank.account.persistence.repository;
+package com.fisa.bank.account.persistence.repository.jpa;
 
 import java.util.Optional;
